refactor(governance): 拆分 daemon 线程标题生成模块
